Clean up database selection implementation
